client5: delete WalletViewModelFactory and replace it with AnyViewModelFactory, also moved wallet models in package
