+++ RichFamily/app/src/main/java/ru/vsu/cs/tp/richfamily/api/model/wallet/WalletRequestBody.kt
@@ -1,4 +1,4 @@
-package ru.vsu.cs.tp.richfamily.api.model
+package ru.vsu.cs.tp.richfamily.api.model.wallet
 
 data class WalletRequestBody(
     val user: Int,
