+++ RichFamily/app/src/main/java/ru/vsu/cs/tp/richfamily/viewmodel/factory/AnyViewModelFactory.kt
@@ -3,20 +3,27 @@ package ru.vsu.cs.tp.richfamily.viewmodel.factory
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import ru.vsu.cs.tp.richfamily.repository.CategoryRepository
+import ru.vsu.cs.tp.richfamily.repository.WalletRepository
 import ru.vsu.cs.tp.richfamily.viewmodel.CategoryViewModel
+import ru.vsu.cs.tp.richfamily.viewmodel.WalletViewModel
 
 class AnyViewModelFactory(
     private val repository: Any,
     private val token: String
 ) : ViewModelProvider.Factory {
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        return if (modelClass.isAssignableFrom(CategoryViewModel::class.java)) {
-            CategoryViewModel(
-                categoryRepository = repository as CategoryRepository,
-                token = token
-            ) as T
-        } else {
-            throw IllegalArgumentException("ViewModel Not Found")
+        return when {
+            modelClass.isAssignableFrom(CategoryViewModel::class.java) ->
+                CategoryViewModel(
+                    categoryRepository = repository as CategoryRepository,
+                    token = token
+                ) as T
+            modelClass.isAssignableFrom(WalletViewModel::class.java) ->
+                WalletViewModel(
+                    walletRepository = repository as WalletRepository,
+                    token = token
+                ) as T
+            else -> throw IllegalArgumentException("ViewModel Not Found")
         }
     }
 }
